// src/app/modules/stock-overview/models/meter.model.ts
export class Meter {
    tehaLgNo?: any;
    adminLgNo?: any;
    tehaUserNo?: any;
    adminUserNo?: any;
    flatId: number;
    tenantName: string;
    meterTypeDesc?: any;
    serialNumber?: any;
    location: string;
    rwmStatus?: any;
    plumbingStatus?: any;
    testType?: any;
    testDate?: any;
    comments?: any;
}